# ENH: faster contains_xy/intersects_xy predicates special casing for point coordinates