--- a/src/ufuncs.c
+++ b/src/ufuncs.c
@@ -390,7 +390,7 @@ static void* intersects_xy_data[1] = {GEOSPreparedIntersectsXY_r};
 static void* contains_xy_data[1] = {GEOSPreparedContains_r};
 static void* intersects_xy_data[1] = {GEOSPreparedIntersects_r};
 #endif
-typedef char FuncGEOS_Ydd_b(void* context, void* g, void* pg, double x, double y);
+typedef char FuncGEOS_Ydd_b(void* context, void* pg, double x, double y);
 static char Ydd_b_p_dtypes[4] = {NPY_OBJECT, NPY_DOUBLE, NPY_DOUBLE, NPY_BOOL};
 static void Ydd_b_p_func(char** args, npy_intp* dimensions, npy_intp* steps, void* data) {
 #if GEOS_SINCE_3_12_0