See https://github.com/libgeos/geos/pull/677: GEOS added new XY versions of some predicates that avoid the overhead of point construction, and started experimenting a bit with those. 
This would also close https://github.com/shapely/shapely/issues/1521

Some points:

* For now I went with an additional `_xy` top-level function (instead of for example integrating it in the existing predicate functions). Are we OK with such an addition to the API?
  * This can also replace the equivalent functions in `shapely.vectorized` (after deprecating them). Also here I think functions with a different name are better than the current `shapely.vectorized.intersects(geom, x, y)` (same name as top-level one, but with a different signature)
* If the geometry is not yet prepared, I followed how we do it elsewhere and prepare/destroy_prepared on the fly. However, given the difference in speed in case you are repeatedly testing against a single geometry (as in the benchmarks below, around 600ms vs 60ms), I am wondering if we shouldn't just document that those functions prepare the geometries (or alternatively, we could also call prepare/destroy_prepare in the python wrapper, avoiding that we prepare the geometry repeatedly in the C function).
* Instead of only making this function available for GEOS 3.12+, I just implemented the "create point + prepared predicate" for older versions of GEOS. This still creates a GEOSGeometry Point, but already avoids creating the Python geometry objects, and based on the timings, this is most of the overhead. So also for older GEOS versions, this specialized `_xy` can give a speedup.

If we want to add this, still need to add tests, expand docstrings, etc.

Some small benchmarks:

```python
import shapely
rng = np.random.default_rng(42)
# use more complex polygon than `polygon = shapely.box(0, 0, 1, 1)`
polygon = shapely.make_valid(shapely.unary_union(shapely.polygons(rng.random((100, 3, 2)))))
x = rng.standard_normal(1_000_000) 
y = rng.standard_normal(1_000_000) 
```

Scalar (I included the change from https://github.com/shapely/shapely/pull/1547 while running those, since that speeds up `Point(..)`):

```
In [2]: %%timeit
   ...: shapely.contains_xy(polygon, 0.5, 0.5)
8.87 µs ± 69.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [3]: %%timeit
   ...: p = shapely.Point(0.5, 0.5)
   ...: shapely.contains(polygon, p)
78.7 µs ± 2.12 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

Scalar with polygon prepared in advance:
```
In [4]: shapely.prepare(polygon)

In [5]: %%timeit
   ...: shapely.contains_xy(polygon, 0.5, 0.5)
3.88 µs ± 66.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [6]: %%timeit
   ...: p = shapely.Point(0.5, 0.5)
   ...: shapely.contains(polygon, p)
9.34 µs ± 157 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

Vectorized:
```
In [6]: %%timeit 
   ...: shapely.contains_xy(polygon, x, y) 
650 ms ± 27.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   # <- GEOS 3.12

In [7]: %%timeit 
   ...: points = shapely.points(x, y) 
   ...: shapely.contains(polygon, points)
8.19 s ± 752 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

(the big difference here above is that I made `contains_xy` to always prepare on the fly, while `contains` does not do this, so this is not a fully fair comparison)

Vectorized with polygon prepared in advance:
```
In [8]: shapely.prepare(polygon)                                                                                                                                                                                   

In [9]: %%timeit 
   ...: shapely.contains_xy(polygon, x, y)
68.6 ms ± 953 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <-- GEOS 3.11
47.8 ms ± 1.44 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <-- GEOS 3.12 dev

In [10]: %%timeit 
    ...: points = shapely.points(x, y) 
    ...: shapely.contains(polygon, points)
356 ms ± 37 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

So it takes more time to create the points compared to just the contains check. To make this explicit:

```
In [7]: %timeit points = shapely.points(x, y)
299 ms ± 12 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

points = shapely.points(x, y)
In [10]: %timeit shapely.contains(polygon, points)
49.1 ms ± 2.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```