@caspervdw 2022-10-06
Looks good @jorisvandenbossche ! 

A high-level comment: can we accept a tuple of coordinates in the existing `contains` function instead of exposing another function? Like:

```shapely.contains(geometry, (0.1, 0.1))``` ?

These could be vectorized as well as a `(i),(i,2)->(i)` ufunc.