@jorisvandenbossche 2022-10-06
Trying to reuse the existing funtions is certainly an option as well (if we can do that cleanly, that would limit a proliferation of `_xy` versions if GEOS adds more of those in the future). 

The main reason I was thinking it doesn't really fit nicely is because I wanted to have separate x and y arguments, and so that would make the signature not very clean (something like `contains(a, b, y=None)` with `b` getting interpreted as `x` if `y` is also specified ...).

Passing xy as a single (n, 2) array (or a tuple as the scalar (1D array) version of that) would make it indeed fit better in the existing function. We would then need to dispatch under the hood to the actual ufunc depending on the data type of the second argument, I suppose? (since the shape itself can't be relied upon, given that also for geometries you could have such a shape with the broadcasting abilities of numpy) 
Although I would certainly be OK with such an API, there is a significant difference of passing separate arrays or a single array. If you start with separate arrays for x and y (for example, if you represent points as x and y columns in a dataframe), then you first need to interleave those (horizontally stack) for being able to pass them as a (n, 2) array.