@caspervdw 2022-10-07
> Although I would certainly be OK with such an API, there is a significant difference of passing separate arrays or a single array. If you start with separate arrays for x and y (for example, if you represent points as x and y columns in a dataframe), then you first need to interleave those (horizontally stack) for being able to pass them as a (n, 2) array.

There is indeed a difference when working with arrays. But that goes both ways; if you happen to start with an `(n, 2)` you would have to transpose and unpack them. Do you have a feeling of what users tend to start out with?
My intuitive preference would be `(n, 2)` because the two numbers (x and y) that form 1 conceptual thing  (a Point) are together.