@jorisvandenbossche 2022-10-07
If you have a (n, 2) array, you can still easily slice and pass x and y as separate arguments (they have just a stride of x2), and performance wise that doesn't change much compared to separate x and y arrays. Using the variables from above:

```
In [9]: xy = np.stack((x, y), axis=-1)

In [10]: %timeit np.stack((x, y), axis=-1)
3.81 ms ± 521 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [11]: shapely.prepare(polygon)

In [12]: %timeit shapely.contains_xy(polygon, x, y)
69.1 ms ± 388 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [13]: %timeit shapely.contains_xy(polygon, xy[:, 0], xy[:, 1])
69.5 ms ± 356 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

While the other way around you need to concatenate (and thus copy) the data. Although, given the above timing of that step, this copying is also not that significant compared to the actual operation. So this is probably not a strong argument either way :)