@jorisvandenbossche 2022-10-24
To follow-up on this: for the actual ufunc, we will certainly keep two separate implementations (since they have different data types and signature), although I might change the new ufunc to a generalized ufunc with reducing dimension for xy (as Casper mentioned above: https://github.com/shapely/shapely/pull/1548#issuecomment-1270509138).  
But so the question is what to do with the public API. 

**Option 1**: expose those as separate functions (current state of the PR), meaning we have both a `shapely.contains(geom1, geom2)` and `shapely.contains_xy(geom1, xy)` (or `shapely.contains_xy(geom1, x, y)`, see below). And same for `intersects`.

**Option 2**: expose the new functionality to pass points as xy coordinates in the existing `contains` function. We would then dispatch to the underlying ufunc based on the data type, roughly like:

```python
def contains(geom1, geom2):
    # ... coerce to arrays
    if np.issubdtype(geom2.dtype, np.number):
        # geom2 are float coordinates
        return lib.contains_xy(geom1, geom2)
    else:
        return lib.contains(geom1, geom2)
```

Some considerations:

* Pros/cons of both options:
  * Re-using the existing functions avoids creating new functions for essentially the same operation (only a different way to express the second argument). This could also be a more scalable pattern if we want to add this ability to more functions. 
    I think this is the biggest plus for combining them in a single function. 
  * For discoverability, I think one could argue both ways. Having an explicit `contains_xy` in the namespace (which you for example see when tab-completing) might make this more visible. On the other hand adding this to the existing `contains` and describing this option in that docstring, might make it easier to see for the many users of that function.
  * A slight disadvantage of combining both in a single function is that it makes the input / output shape logic depending on the data type (`(.., n), (.., n) -> (.., n)` for objects vs `(.., n), (.., n, 2) -> (.., n)` for floats)
* In case of doing it as a separate function (option 1), we need to decide whether we want to pass the coordinates as a single xy interleaved array (`contains_xy(geom1, xy)`) or as two separate x, y arrays (`contains_xy(geom1, x, y)`.
  * The current PR does the second (separate x, y) arrays, but that can easily be changed. 
  * The existing [`shapely.vectorized.contains`](https://github.com/shapely/shapely/blob/8c66df52c9c2e04594cc7226708fbd8d7ba7722a/shapely/vectorized/__init__.py#L22) uses separate x, y arrays. On the other hand, our functions that work with coordinates (eg `get_coordinates`) typically work with interleaved xy arrays, so that would be more consistent with that.
  * If you have a single interleaved array, it is easy to slice and pass as separate arrays, while if you have two separate arrays, you need to stack (i.e. copy) the arrays to be able to pass it as a single array. Although the time to make this copy is also quite small compared to the actual operation, so this is not too important (xref [#1548 (comment)](https://github.com/shapely/shapely/pull/1548#issuecomment-1271999659) above).
  * If we go for a single combined function (option 2), I think we want a single interleaved xy array anyway, since we want to keep the number of arguments the same in both cases (so the second argument is always either an object dtype array of geometries or a float array of xy coordinates). 

cc @caspervdw @brendan-ward @sgillies @mwtoews preferences / thoughts?