@brendan-ward 2022-10-27
We discussed this a little bit during the GeoPandas dev meeting.  While there is some benefit of clarity to having separate methods to keep the documentation around the _xy predicates very clear, there is already a precedent of functions in Shapely 2.0 that take Points or coordinates as input: the `LineString` constructor takes both.  

Thus it seems like the single combined option with interleaved xy array seems like the most straightforward way to add this.