@jorisvandenbossche 2022-10-28
>  there is already a precedent of functions in Shapely 2.0 that take Points or coordinates as input: the LineString constructor takes both.

This is only for the `LineString()` class constructor though (shapely already had this behaviour for a long time), and not for the `shapely.linestrings(..)` bulk construction function.